refactor: MediaAttachment 파일 분리
