+++ MemorialHouse/MHPresentation/MHPresentation/Source/EditBook/View/MediaAttachable.swift
@@ -1,0 +1,7 @@
+import MHFoundation
+import MHDomain
+
+protocol MediaAttachable {
+    func configureSource(with mediaDescription: MediaDescription, data: Data)
+    func configureSource(with mediaDescription: MediaDescription, url: URL)
+}
